requirements.txt
fastapi
uvicorn
httpx
openai
python-dotenv
pydantic
streamlit
